community_guardian/requirements.txt
python-dotenv==1.0.1
pytest==8.3.5
streamlit==1.38.0
